create useChartPan hook
