+++ client/src/Components/charts/OrgChart.js
@@ -1,4 +1,4 @@
-import React, { Fragment, useRef, useEffect, useState } from "react";
+import React, { Fragment, useRef } from "react";
 import OrganizationChart from "@dabeng/react-orgchart";
 import OrgChartNode from "./OrgChartNode";
 
@@ -8,39 +8,6 @@ import "./OrgChart.scss";
 
 const OrgChart = () => {
   const orgChartRef = useRef(null);
-  const orgChartElRef = useRef(null);
-
-  const [translateX, setTranslateX] = useState("0");
-  const [translateY, setTranslateY] = useState("0");
-
-  useEffect(() => {
-    // Hook onto orgchart DOM element
-    orgChartElRef.current = document.querySelector(".orgchart.myChart");
-  }, []);
-
-  useEffect(() => {
-    const transformValue = orgChartElRef.current.style.transform;
-    if (transformValue.includes("translateX")) {
-      orgChartElRef.current.style.transform = transformValue.replace(
-        /translateX\([0-9\.px\-]+\)/,
-        translateX
-      );
-    } else {
-      orgChartElRef.current.style.transform = `${transformValue} ${translateX}`;
-    }
-  }, [translateX]);
-
-  useEffect(() => {
-    const transformValue = orgChartElRef.current.style.transform;
-    if (transformValue.includes("translateY")) {
-      orgChartElRef.current.style.transform = transformValue.replace(
-        /translateY\([0-9\.px\-]+\)/,
-        translateY
-      );
-    } else {
-      orgChartElRef.current.style.transform = `${transformValue} ${translateY}`;
-    }
-  }, [translateY]);
 
   const handleDownload = async () => {
     orgChartRef.current.exportTo("chart", "png");
@@ -124,7 +91,7 @@ const OrgChart = () => {
   return (
     <Fragment>
       <ZoomControl />
-      <PanControl setTranslateX={setTranslateX} setTranslateY={setTranslateY} />
+      <PanControl />
       <OrganizationChart
         ref={orgChartRef}
         datasource={ds}
